Never Expire

+++ app.py
@@ -12,6 +12,7 @@ snippets = {}
 
 # "precomputed" durations in seconds. 5 minutes = 5 * 60 etc.
 DURATION_OPTIONS = {
+    'Never Expire': -1,
     '5 minutes': 300,
     '10 minutes': 600,
     '30 minutes': 1800,
@@ -41,7 +42,10 @@ async def save_snippet():
     snippet_name = form['snippetName']
     duration = int(form.get('duration', 0))
     url = generate_url()
-    expiration_time = time.time() + duration
+    if duration == -1:
+        expiration_time = -1
+    else:
+        expiration_time = time.time() + duration
 
     if not snippet_name.strip():
         error_message = 'The Snippet\'s name must not be empty!'
@@ -52,30 +56,31 @@ async def save_snippet():
 
     try:
         conn.execute("INSERT INTO snippets (url, snippet, snippet_name, expires_at) VALUES (?, ?, ?, ?)",
-                 (url, snippet, snippet_name, expiration_time))
+                     (url, snippet, snippet_name, expiration_time))
         conn.commit()
     except sqlite3.OperationalError as e:
         print(f"{e}\nCould not find \"snippets.db\"")
-        return jsonify({'error': 'Internal Error: Database not found. Possible causes: Database is deleted or renamed.'})
-    return redirect(f'/snippet/{url}')
+        return jsonify(
+            {'error': 'Internal Error: Database not found. Possible causes: Database is deleted or renamed.'})
 
+    return redirect(f'/snippet/{url}')
 
 
 @app.route('/snippet/<url>')
 async def get_snippet(url):
-    cursor = conn.execute("SELECT snippet, snippet_name, expires_at FROM snippets WHERE url = ?", (url,))
-    row = cursor.fetchone()
-    if row is not None:
-        snippet_data = {'snippet': row[0], 'snippet_name': row[1], 'expires_at': row[2]}
-        if time.time() > snippet_data['expires_at']:
-            conn.execute("DELETE FROM snippets WHERE url = ?", (url,))
-            conn.commit()
-            return "Snippet has expired!"
-        snippet = snippet_data['snippet']
-        snippet_name = snippet_data['snippet_name']
-        return await render_template('snippet.html', snippet=snippet, snippet_name=snippet_name)
-    else:
-        return "Snippet not found!"
+  cursor = conn.execute("SELECT snippet, snippet_name, expires_at FROM snippets WHERE url = ?", (url,))
+  row = cursor.fetchone()
+  if row is not None:
+    snippet_data = {'snippet': row[0], 'snippet_name': row[1], 'expires_at': row[2]}
+    if snippet_data['expires_at'] != -1 and time.time() > snippet_data['expires_at']:
+      conn.execute("DELETE FROM snippets WHERE url = ?", (url,))
+      conn.commit()
+      return "Snippet has expired!"
+    snippet = snippet_data['snippet']
+    snippet_name = snippet_data['snippet_name']
+    return await render_template('snippet.html', snippet=snippet, snippet_name=snippet_name)
+  else:
+    return "Snippet not found!"
 
 
 if __name__ == '__main__':
